requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
colorlog==4.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.1
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
har2case==0.3.1
HttpRunner==1.5.8
idna==2.9
itypes==1.1.0
Jinja2==2.11.1
Markdown==3.2.1
MarkupSafe==1.1.1
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-toolbelt==0.9.1
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.8
